wip: move 'rtc' folder into common pkg
