+++ packages/common/src/webrtc/Dialog.ts
@@ -1,16 +1,16 @@
 import { v4 as uuidv4 } from 'uuid'
-import logger from '../../../common/src/util/logger'
+import logger from '../util/logger'
 import BrowserSession from '../BrowserSession'
-import BaseMessage from '../../../common/src/messages/BaseMessage'
-import { Invite, Answer, Attach, Bye, Modify, Info } from '../../../common/src/messages/Verto'
+import BaseMessage from '../messages/BaseMessage'
+import { Invite, Answer, Attach, Bye, Modify, Info } from '../messages/Verto'
 import Peer from './Peer'
-import { PeerType, VertoMethod, SwEvent, NOTIFICATION_TYPE, Direction } from '../../../common/src/util/constants'
-import { State, DEFAULT_DIALOG_OPTIONS, ConferenceAction, Role } from '../../../common/src/util/constants/dialog'
-import { trigger, register, deRegister } from '../../../common/src/services/Handler'
+import { PeerType, VertoMethod, SwEvent, NOTIFICATION_TYPE, Direction } from '../util/constants'
+import { State, DEFAULT_DIALOG_OPTIONS, ConferenceAction, Role } from '../util/constants/dialog'
+import { trigger, register, deRegister } from '../services/Handler'
 import { streamIsValid, sdpStereoHack, sdpMediaOrderHack, checkSubscribeResponse } from './helpers'
-import { objEmpty, mutateLiveArrayData, isFunction } from '../../../common/src/util/helpers'
-import { attachMediaStream, detachMediaStream } from '../../../common/src/util/webrtc'
-import { DialogOptions } from '../../../common/src/util/interfaces'
+import { objEmpty, mutateLiveArrayData, isFunction } from '../util/helpers'
+import { attachMediaStream, detachMediaStream } from './utils'
+import { DialogOptions } from '../util/interfaces'
 
 export default class Dialog {
   public id: string = ''
