+++ packages/common/src/webrtc/utils.ts
@@ -1,10 +1,8 @@
-import logger from '../logger'
-import { findElementByType } from '../helpers'
+import { findElementByType } from '../util/helpers'
 
 const attachMediaStream = (tag: any, stream: MediaStream) => {
   const element = findElementByType(tag)
   if (element === null) {
-    // logger.warn(`Unknown HTML element for ${tag}.`)
     return
   }
   if (!element.getAttribute('autoplay')) {
